requirements.txt
torch
torchvision
captcha
pillow
numpy
